fix(index.jelly): selected value is now used (#2)
* fix(index.jelly): selected value is now used

This fixes an issue where selected value is not used when defaultValue was empty. Changing the if-statement fixed the issue. And disabled jenkins-security-scan.yml in this ticket too.

Closes #1

* fix(index.jelly): selected value is now used

This fixes an issue where selected value is not used when defaultValue was empty. Changing the if-statement fixed the issue. And disabled jenkins-security-scan.yml in this ticket too.

Closes #1